+++ Changes.md
@@ -9,6 +9,7 @@
 - Removed the 'started' worker data, it originally provided compatibility with resque-web
   but overlaps the 'run_at' worker data.
 - Remove built-in error integration for Airbrake, Honeybadger, ExceptionNotifier and Exceptional
+- Remove Sidekiq::Client.registered\_\* APIs